🚀 feat(_app.tsx): add Google Fonts to the application This commit adds Google Fonts to the application using the next-google-fonts package. Two fonts, DM Sans and Syne, are imported and used in the application.
🎨 style(global.tsx): update global styles
This commit updates the global styles of the application. The primary color has been changed to #03DAC5, and a primary variant color has been added. The primary glass and primary glass variant have been added to the global styles. The secondary color has been changed to #BB86FC, and a secondary variant color has been added. The gray color palette has been updated with new shades. Box shadow variables have been added to the global styles. A background blur variable has been added to the global styles. The font sizes have been updated, and new font size variables have been added. The font weight variables have been updated. Spacing variables have been added to the global styles.

🎨 style(global.tsx): update global styles
This commit updates the global styles of the application. The font family of the application has been updated to 'DM Sans' for better readability. The background color of the application has been changed to a darker shade of gray, 'var(--gray-900)'. The font style of the text has been set to normal, and the color has been changed to 'var(--gray-50)'. The line-height of the text has been set to 24px, and the text-wrap has been set to balance. The max-width of the section has been changed to 1440px, and the padding has been updated to use custom CSS variables. The font family of the headings has been updated to 'Syne', and the font-weight has been set to bold. The line-height and letter-spacing of the headings have been updated to improve readability.

🎨 style(global.tsx): remove line-height and change font-weight to bold for headings and links
This commit removes the line-height property from h3, h4 and links. It also changes the font-weight property from semibold to bold for links and headings. This improves the readability and consistency of the text across different devices.