🎨 style(expertise, header): update media query for smaller screens The media query for smaller screens has been updated from 768px to 1024px in the expertise component. In the header component, a new media query has been added to adjust the padding of the header container for smaller screens. This improves the responsiveness of the application on smaller screens.
🎨 style(hero): restructure hero component layout
The hero component layout has been restructured to improve the visual hierarchy of the content. The title, subtitle, and button are now in a separate column from the avatar image. This improves the readability of the text and makes the content more visually appealing. The avatar image has also been updated to use the next/image component for better performance.

🎨 style(hero): adjust hero section layout for responsiveness
The hero section has been updated to improve its responsiveness. The Container component has been renamed to Wrapper and now has a flex display with left text alignment and padding. The Title and Subtitle components have been updated to have a width of 100%. The Column component has been added to wrap the image and text content, with a flex display, column direction, and gap. The last child of the Column component has a width of 30% and aligns its items to the right. The ScrollLink component has been updated to have a margin-top of var(--space-md) and a hover effect on the svg. The media query has been added to adjust the padding of the Wrapper component and the Column component's alignment and width when the screen size is less than or equal to 1024px.

🎨 style(styles.tsx): update media query breakpoints
The media query breakpoints for the Projects and Skills components have been updated to use a max-width of 1024px instead of 768px. This change improves the responsiveness of the components on larger screens.

🔥 refactor(global.tsx): remove media queries for font sizes
🎨 style(global.tsx): simplify font declaration for headings
The media queries for font sizes have been removed as they were not necessary and made the code harder to read. The font declaration for headings has been simplified by using the shorthand font property.

🎨 style(global.tsx): add media queries for responsive font sizes
This commit adds media queries to adjust font sizes for different screen sizes. The font sizes are now responsive and will adjust based on the screen size. The media queries are added for screens with a maximum width of 1024px and 600px.