🎨 style(about/index.tsx): refactor About component to use new styles 🚀 chore(about/index.tsx): replace AboutIllustration with AboutImage The About component has been refactored to use new styles. The AboutIllustration component has been replaced with the AboutImage component, which is an SVG image. This change improves the performance of the component as SVG images are smaller in size and can be scaled without losing quality.
🎨 style(about/styles.tsx): change AboutFlexContainer to Wrapper, change svg to img, add flex-direction and width properties
🎨 style(about/styles.tsx): change AboutSectionTitle to Title, remove text-transform property
🎨 style(about/styles.tsx): change AboutDescription to Description
🎨 style(button/styles.tsx): add width property to ButtonStyles
The changes in the about/styles.tsx file include renaming the AboutFlexContainer to Wrapper, changing the svg element to img, and adding the flex-direction and width properties. The AboutSectionTitle was renamed to Title and the text-transform property was removed. The AboutDescription was renamed to Description. In the button/styles.tsx file, the width property was added to the ButtonStyles. These changes improve the consistency and readability of the code.