🎉 feat(experience): add experience data component This commit adds a new component that exports an array of objects containing information about work experiences. Each object contains the company logo, name, details, position, and a list of responsibilities. This component will be used to display work experience information on the website.
🎨 style(experience): refactor Experience component to use styled components
✨ feat(experience): add accordion functionality to display experience details
The Experience component has been refactored to use styled components, which improves readability and maintainability. The component now also includes an accordion functionality that allows the user to toggle the display of experience details. This functionality has been implemented using the useState hook and the open state variable. The toggleAccordion function updates the open state variable to reflect the new state of the accordion. The Experiences data is now imported from a separate file.

✨ feat(experience): add styles for experience component
This commit adds styles for the experience component. The styles include a section, title, wrapper, accordion, header, container, company, details, content, position, intro, responsibilities, and responsibility. The styles are used to create a visually appealing experience component with a gradient background, box-shadow, and a toggle icon that rotates when the accordion is opened.

🎨 style(header, global): remove width: 100% from NavContainer and add transition-03 variable to global styles
💄 Improve consistency by removing the width: 100% property from NavContainer in the header component. Add a new variable to global styles, --transition-03, which will be used for all transitions in the application.