🔄 refactor(header): update import path for ThemeContext 🔧 chore(header): add sticky header functionality The import path for the ThemeContext has been updated to reflect the new location of the file. Additionally, a new state variable `isSticky` has been added to handle the sticky header functionality. The `isSticky` state is updated based on the scroll position of the window, and the header container now receives a `issticky` attribute to apply styling based on the sticky state.
🔧 chore(menu-button): import dependencies and add audio click functionality
🔧 chore(nav-link): refactor NavLink component to use Button component
🔧 chore(nav-links): import AudioToggle component
The MenuButton component now imports the AudioContext and PlayAudio components from their respective paths. It also uses the useContext hook to access the audio state from the AudioContext. The onClick event of the MenuButton now toggles the menu and plays an audio click sound if the audio state is set to 'on'.

The NavLink component has been refactored to use the Button component instead of the LinkStyles component. This change allows for better consistency and reusability of the button styling.

The NavLinks component now imports the AudioToggle component and includes it in the rendered JSX. This allows for the toggling of audio functionality within the header navigation links.

🔧 chore(header/styles.tsx): refactor HeaderContainer styles
🌈 style(header/styles.tsx): add box-shadow and transition to HeaderContainer

The HeaderContainer component's styles have been refactored to accept an issticky prop, which determines whether the position should be sticky or unset. Additionally, a box-shadow and transition have been added to the HeaderContainer component to enhance the visual appearance.