🔧 fix(styles.tsx): remove transition property from Summary and CompanyLogo components 🔧 fix(header/index.tsx): add explicit types to useState hooks and adjust scroll threshold The transition property has been removed from the Summary and CompanyLogo components in order to improve performance and eliminate unnecessary animations. In the header/index.tsx file, explicit types have been added to the useState hooks to improve code clarity. Additionally, the scroll threshold for the isSticky state has been adjusted to 81 pixels to ensure the header becomes sticky at the desired position.
🐛 fix(header/styles.tsx): fix typo in isSticky prop name
🐛 fix(hero/styles.tsx): fix typo in min-height value

In the header/styles.tsx file, there was a typo in the prop name isSticky, which caused the position and box-shadow properties to not be applied correctly. This has been fixed by changing the prop name to isSticky.

In the hero/styles.tsx file, there was a typo in the min-height value, which caused the height of the section to be incorrect. This has been fixed by changing the value from 100dvh to 100vh.

🐛 fix(scroll-to-top): specify type for isVisible state variable
🔧 chore(scroll-to-top): update ScrollToTopButton component props type
🔧 chore(globalStyles): remove unnecessary transition property

The `isVisible` state variable in the `ScrollToTop` component now has an explicit type of `boolean` to improve type safety. The `ScrollToTopButton` component's `visible` prop has been updated to accept a `boolean` value instead of a string. Additionally, the unnecessary `transition` property has been removed from the global styles.